the SavePtr union in sv_save.c doesn't work correctly on big endian, at least not on the remote HPUX host I tested. so, here it is: changed SavePtr to be a byte* pointer and made GET_BYTE, GET_WORD and GET_LONG macros into inline functions which increment the pointer properly after reading. the functions read the data in an endian-neutral way, too, and gcc will output good optimized code for them, I think. short and int values are now written in little endian format. the missing thing is that we don't properly byte swap the structures we are reading and writing through StreamOutBuffer: it will be done later.  while I was there, made a similar change in p_acs.c for PCodePtr: PCodePtr is a byte pointer now, instead of an int pointer. it is read in an endian-neutral way bythe newly added inline functions ReadCodePtr, IncrCodePtr and ReadAndIncrCodePtr.
git-svn-id: svn://svn.code.sf.net/p/hhexen/code/trunk/hhexen@205 989aa981-fe4f-0410-aa6a-c19d413bcdcd

+++ include/p_spec.h
@@ -4,8 +4,8 @@
 //** p_spec.h : Heretic 2 : Raven Software, Corp.
 //**
 //** $RCSfile: p_spec.h,v $
-//** $Revision: 1.6 $
-//** $Date: 2008-06-22 16:32:45 $
+//** $Revision: 1.7 $
+//** $Date: 2008-07-07 20:25:47 $
 //** $Author: sezero $
 //**
 //**************************************************************************
@@ -481,7 +481,7 @@ typedef struct acsInfo_s acsInfo_t;
 struct acsInfo_s
 {
 	int		number;
-	int		*address;
+	byte		*address;
 	int		argCount;
 	aste_t		state;
 	int		waitValue;
@@ -499,7 +499,7 @@ struct acs_s
 	int		stack[ACS_STACK_DEPTH];
 	int		stackPtr;
 	int		vars[MAX_ACS_SCRIPT_VARS];
-	int		*ip;
+	byte		*ip;
 };
 
 typedef struct
